Moved SearchError to error.py
