+++ src/arguments/utility.py
@@ -2,28 +2,12 @@ import requests
 from clint.textui import colored as TextColor
 from rich.console import Console
 from html import unescape
-from .markdown import MarkdownRenderer
 import sys as sys
 
+from .markdown import MarkdownRenderer
+from .error import SearchError
+
 console = Console()
-
-class SearchError():
-    def __init__(self, error_statement, suggestion="Try again"):
-        # the error statement
-        self.error_statement = error_statement
-
-        # the suggestion statement
-        self.suggestion = suggestion
-
-        self.evoke_search_error(self.error_statement)
-
-    def evoke_search_error(self, error_statement):
-        print_text = [
-            TextColor.red(error_statement),
-            TextColor.green(self.suggestion)
-        ]
-        for text_to_print in print_text:
-            print(text_to_print)
 
 
 class Utility():
